# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Team Photo Album API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/registration-settings:
    patch:
      operationId: updateRegistrationSettings
      tags: [admin]
      summary: Update registration settings (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegistrationSettingsUpdate"
      responses:
        "200":
          description: Updated registration settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegistrationSettings"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
  /admin/ai-settings:
    get:
      operationId: getAiSettings
      tags: [admin]
      summary: Get AI provider settings (admin only)
      responses:
        "200":
          description: Current AI settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiSettings"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
    patch:
      operationId: updateAiSettings
      tags: [admin]
      summary: Update AI enabled flag and active provider (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiSettingsUpdate"
      responses:
        "200":
          description: Updated AI settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiSettings"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
  /admin/ai-settings/providers/{provider}/key:
    put:
      operationId: setAiProviderKey
      tags: [admin]
      summary: Save (or replace) the API key for a provider (admin only)
      parameters:
        - name: provider
          in: path
          required: true
          schema:
            type: string
            enum: [openai, anthropic, gemini]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiProviderKeyInput"
      responses:
        "200":
          description: Updated AI settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiSettings"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
    delete:
      operationId: clearAiProviderKey
      tags: [admin]
      summary: Clear the saved API key for a provider (admin only)
      parameters:
        - name: provider
          in: path
          required: true
          schema:
            type: string
            enum: [openai, anthropic, gemini]
      responses:
        "200":
          description: Updated AI settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiSettings"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
  /admin/ai-analysis-events:
    get:
      operationId: listAiAnalysisEvents
      tags: [admin]
      summary: List recent AI photo analysis events (admin only)
      responses:
        "200":
          description: Recent AI analysis events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AiAnalysisEvent"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
  /admin/ai-analysis-events/retry-all:
    post:
      operationId: bulkRetryAiAnalysisEvents
      tags: [admin]
      summary: Re-run AI analysis for all recently failed events (admin only)
      responses:
        "200":
          description: Bulk retry result summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BulkRetryAiAnalysisEventsResult"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
  /admin/ai-analysis-events/{id}/retry:
    post:
      operationId: retryAiAnalysisEvent
      tags: [admin]
      summary: Re-run AI analysis for the photo of a failed event (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: New AI analysis event recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiAnalysisEvent"
        "400":
          description: Event is not in a retryable state
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Event or its photo not found
...
components:
  schemas:
...
    RegistrationSettings:
      type: object
      required: [registrationEnabled]
      properties:
        registrationEnabled:
          type: boolean
    RegistrationSettingsUpdate:
      type: object
      required: [registrationEnabled]
      properties:
        registrationEnabled:
          type: boolean
    AiModelOption:
      type: object
      required: [id, label]
      properties:
        id:
          type: string
        label:
          type: string
    AiProviderInfo:
      type: object
      required: [id, label, model, availableModels, hasKey, replitFallbackAvailable, usable]
      properties:
        id:
          type: string
          enum: [openai, anthropic, gemini]
        label:
          type: string
        model:
          type: string
        availableModels:
          type: array
          items:
            $ref: "#/components/schemas/AiModelOption"
        hasKey:
          type: boolean
        keyPreview:
          type: ["string", "null"]
        replitFallbackAvailable:
          type: boolean
        usable:
          type: boolean
    AiSettings:
      type: object
      required: [enabled, activeProvider, providers, hasUsableActive]
      properties:
        enabled:
          type: boolean
        activeProvider:
          type: string
          enum: [openai, anthropic, gemini]
        hasUsableActive:
          type: boolean
        providers:
          type: object
          required: [openai, anthropic, gemini]
          properties:
            openai:
              $ref: "#/components/schemas/AiProviderInfo"
            anthropic:
              $ref: "#/components/schemas/AiProviderInfo"
            gemini:
              $ref: "#/components/schemas/AiProviderInfo"
    AiSettingsUpdate:
      type: object
      properties:
        enabled:
          type: boolean
        activeProvider:
          type: string
          enum: [openai, anthropic, gemini]
        providerModels:
          type: object
          properties:
            openai:
              type: string
            anthropic:
              type: string
            gemini:
              type: string
    AiProviderKeyInput:
      type: object
      required: [apiKey]
      properties:
        apiKey:
          type: string
          minLength: 4
...
    BulkRetryAiAnalysisEventsResult:
      type: object
      required: [succeeded, skipped, failed]
      properties:
        succeeded:
          type: integer
        skipped:
          type: integer
        failed:
          type: integer
    AiAnalysisEvent:
      type: object
      required: [id, status, createdAt]
      properties:
        id:
          type: integer
        photoId:
          type: ["integer", "null"]
        photoCaption:
          type: ["string", "null"]
        photoThumbnailUrl:
          type: ["string", "null"]
        provider:
          type: ["string", "null"]
        status:
          type: string
          enum: [success, skipped, failed]
        errorMessage:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time